+++ src/app/api/admin/schedule/[id]/route.ts
@@ -1,4 +1,5 @@
 import schedule from "@/model/schedule.models";
+import User from "@/model/user.models";
 import getUser from "@/utils/getUserFromCookie";
 import { NextRequest, NextResponse } from "next/server";
 
@@ -14,6 +15,15 @@ export const DELETE = async (req: NextRequest, res: NextResponse) => {
         success: false,
       });
     }
+
+    const Admin = await User.findById(user);
+    if (!Admin.isAdmin) {
+      return NextResponse.json({
+        message: "You are not authorized to create student",
+        status: 401,
+      });
+    }
+
     const deletedSchedule = await schedule.findByIdAndDelete(id);
     return NextResponse.json({
       message: "Schedule deleted successfully",
